+++ rpc/endorsement.go
@@ -13,10 +13,12 @@ var _ TypedOperation = (*Endorsement)(nil)
 // Endorsement represents an endorsement operation
 type Endorsement struct {
 	Generic
-	Level       int64               `json:"level"`       // <= v008
-	Metadata    OperationMetadata   `json:"metadata"`    // all protocols
-	Endorsement *InlinedEndorsement `json:"endorsement"` // v009+
-	Slot        int                 `json:"slot"`        // v009+
+	Level       int64               `json:"level"`                 // <= v008, v012+
+	Metadata    OperationMetadata   `json:"metadata"`              // all protocols
+	Endorsement *InlinedEndorsement `json:"endorsement,omitempty"` // v009+
+	Slot        int                 `json:"slot"`                  // v009+
+	Round       int                 `json:"round"`                 // v012+
+	PayloadHash tezos.PayloadHash   `json:"block_payload_hash"`    // v012+
 }
 
 func (e Endorsement) GetLevel() int64 {